refactor: reorganize appletSerializer into multiple submodules
