+++ ilastik/applets/base/appletSerializer/__init__.py
@@ -1,0 +1,33 @@
+###############################################################################
+#   ilastik: interactive learning and segmentation toolkit
+#
+#       Copyright (C) 2011-2024, the ilastik developers
+#                                <team@ilastik.org>
+#
+# This program is free software; you can redistribute it and/or
+# modify it under the terms of the GNU General Public License
+# as published by the Free Software Foundation; either version 2
+# of the License, or (at your option) any later version.
+#
+# In addition, as a special exception, the copyright holders of
+# ilastik give you permission to combine ilastik with applets,
+# workflows and plugins which are not covered under the GNU
+# General Public License.
+#
+# See the LICENSE file for details. License information is also available
+# on the ilastik web site at:
+#          http://ilastik.org/license.html
+###############################################################################
+from .appletSerializer import AppletSerializer as AppletSerializer
+from .serializerUtils import deleteIfPresent as deleteIfPresent
+from .serializerUtils import getOrCreateGroup as getOrCreateGroup
+from .slotSerializer import JSONSerialSlot as JSONSerialSlot
+from .slotSerializer import SerialBlockSlot as SerialBlockSlot
+from .slotSerializer import SerialClassifierFactorySlot as SerialClassifierFactorySlot
+from .slotSerializer import SerialClassifierSlot as SerialClassifierSlot
+from .slotSerializer import SerialCountingSlot as SerialCountingSlot
+from .slotSerializer import SerialDictSlot as SerialDictSlot
+from .slotSerializer import SerialListSlot as SerialListSlot
+from .slotSerializer import SerialObjectFeatureNamesSlot as SerialObjectFeatureNamesSlot
+from .slotSerializer import SerialPickleableSlot as SerialPickleableSlot
+from .slotSerializer import SerialSlot as SerialSlot
